requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
attrs==24.2.0
contourpy==1.3.0
cycler==0.12.1
discord.py==2.4.0
fonttools==4.53.1
frozenlist==1.4.1
idna==3.8
kiwisolver==1.4.7
matplotlib==3.9.2
multidict==6.0.5
numpy==2.1.1
packaging==24.1
pillow==10.4.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
six==1.16.0
ujson==5.10.0
yarl==1.9.8
